requirements.txt
PyYAML
requests
youtube-dl
openai
ffmpeg-python
